Done with refacrtoring of property analyzer.  Build worked.
